Added FileSettingsHash. This class saves the settings of the played file on a different ini file (instead of using only one big ini file for all of them).
The files are saved under the subdirectory "file_settings" in the config
directory (usually $HOME/.config/smplayer/). As filename it uses a hash,
and it's stored on a subdirectory starting with the first letter of the
filename. 

For example:
$HOME/.config/smplayer/file_settings/d/dd017f9d6706a75f.ini
$HOME/.config/smplayer/file_settings/9/95e963e59826699d.ini

The hash for the filenames, for simplicity (it's already coded and it's fast) 
is the same used for opensubtitles.